virustotal-enrich-connector/requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dacite==1.6.0
datefinder==0.7.3
idna==3.4
pathspec==0.10.2
pika==1.3.1
platformdirs==2.5.4
pycti==5.4.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-magic==0.4.27
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sseclient==0.0.27
stix2==3.0.1
stix2-patterns==2.0.0
tomli==2.0.1
urllib3==1.26.13
